Add assertions
Several methods assume that input is in accordance with the length certain lists, or is of certain type.

Adding assertions improves the code quality by reducing the likelihood of that fatal runtime errors occurring.

Let's add these assertions at crucial points in the code that assume limitations on the values of some input parameters.